+++ src/main/java/com/mageddo/dnsproxyserver/docker/NetworkComparator.java
@@ -5,7 +5,7 @@ import com.github.dockerjava.api.model.Network;
 public class NetworkComparator {
 
   static int toPriorityOrder(Network n) {
-    return com.mageddo.dnsproxyserver.docker.Network.of(n.getName()).ordinal();
+    return com.mageddo.dnsproxyserver.docker.domain.Network.of(n.getName()).ordinal();
   }
 
   public static int compare(Network a, Network b) {
