+++ src/main/java/com/mageddo/dnsproxyserver/docker/domain/Network.java
@@ -1,6 +1,7 @@
-package com.mageddo.dnsproxyserver.docker;
+package com.mageddo.dnsproxyserver.docker.domain;
 
 import org.apache.commons.lang3.EnumUtils;
+import org.apache.commons.lang3.StringUtils;
 
 public enum Network {
 
@@ -12,4 +13,7 @@ public enum Network {
     return EnumUtils.getEnumIgnoreCase(Network.class, name, OTHER);
   }
 
+  public String lowerName() {
+    return StringUtils.lowerCase(this.name());
+  }
 }
